working GameApp

+++ RandomGo.java
@@ -19,12 +19,12 @@ public class RandomGo{
     //process
     public void CreateGo() {
 		cGo = (int)(Math.random()*3) + 1; 	//tested and generates the random number between [1, 3]
-		System.out.println("cGo random " + cGo); //testing purposes
-		
+		System.out.println("cGo random " + cGo); //testing purposes	
 	}
     
     //getter
-    public double getCGo(){
-		return cGo;
-	}
+    public int getCGo(){
+	return cGo;
+	} 
+  
 }
